fixup! nervous_system::intergration_tests: add nervous-agent helpers
