+++ rs/nervous_system/integration_tests/src/lib.rs
@@ -12,7 +12,6 @@
 //! crate.
 
 pub mod create_service_nervous_system_builder;
-pub mod nervous_system_agent_helpers;
 pub mod pocket_ic_helpers;
 
 use std::time::Instant;
