+++ rs/nervous_system/integration_tests/tests/sns_upgrade_test_utils.rs
@@ -1,15 +1,13 @@
 use ic_base_types::PrincipalId;
+use ic_nervous_system_agent::helpers::await_with_timeout;
 use ic_nervous_system_common::ONE_MONTH_SECONDS;
 use ic_nervous_system_integration_tests::{
     create_service_nervous_system_builder::CreateServiceNervousSystemBuilder,
     pocket_ic_helpers::{
-        self, await_with_timeout, hash_sns_wasms, nns,
-        sns::{
-            self,
-            governance::{
-                set_automatically_advance_target_version_flag,
-                EXPECTED_UPGRADE_DURATION_MAX_SECONDS, EXPECTED_UPGRADE_STEPS_REFRESH_MAX_SECONDS,
-            },
+        self, hash_sns_wasms, nns, sns,
+        sns::governance::{
+            set_automatically_advance_target_version_flag, EXPECTED_UPGRADE_DURATION_MAX_SECONDS,
+            EXPECTED_UPGRADE_STEPS_REFRESH_MAX_SECONDS,
         },
     },
     SectionTimer,
